TagGpsTimebaseAligner done and collected data modified:   geo_mapping/geo_mapper.py modified:   legacy_data_pipeline/calibration_data_filer.py modified:   legacy_data_pipeline/tag_gps_timebase_aligner.py new file:   tk_canvas_renderers/tag_gps_timebase_aligner_ui.py new file:   tests/test_tag_gps_timebase_aligner.py modified:   tk_canvas_renderers/geo_map_scrubber.py

+++ geo_mapping/geo_mapper.py
@@ -373,11 +373,11 @@ class MapCoordinateTransformer:
 
     def get_y_for_latitude(self, latitude_deg):
         lat_diff = self._get_lat_of_top() - latitude_deg
-        return int(round(lat_diff / self._get_lat_deg_per_px()))
+        return round(lat_diff / self._get_lat_deg_per_px())
 
     def get_x_for_longitude(self, longitude_deg):
         long_diff = longitude_deg - self._get_long_of_left()
-        return int(round(long_diff / self._get_long_deg_per_px()))
+        return round(long_diff / self._get_long_deg_per_px())
 
     def _get_lat_deg_per_px(self):
         if self._lat_deg_per_px is None:
